Add AFX/CSE board support - Kconfig entry, defconfigs
Add RSA Kconfig entry and targets for it
Add zc770-xm011 and zynq-cc108 DTBs from the Linux kernel
Sync zynq-7000.dtsi with the Linux kernel.
Use spl_load_image_fat for bitstream loading in SPL.
Disable CONFIG_OF_CONTROL for all zynq boards.

Nand:
- Fix zynq_nand_read_subpage_raw parameters.
- Remove nand.h from sys_proto.h and move it zynq_nand.c
- Do not call nand_release(). Removed by:
  "mtd, ubi, ubifs: resync with Linux-3.14"
  (sha1: ff94bc40af3481d47546595ba73c136de6af6929)

QSPI:
- spi_flash_params has been moved from spi_flash.h
  to sf_internal.h - fix zynq_qspi.c driver to reflect this.

+++ arch/powerpc/lib/bootm.c
@@ -126,7 +126,7 @@ void arch_lmb_reserve(struct lmb *lmb)
 #endif
 
 	size = min(bootm_size, get_effective_memsize());
-	size = min(size, CONFIG_SYS_LINUX_LOWMEM_MAX_SIZE);
+	size = min(size, (ulong)CONFIG_SYS_LINUX_LOWMEM_MAX_SIZE);
 
 	if (size < bootm_size) {
 		ulong base = bootmap_base + size;
